Add strict-order flag to dnsmasq options, providing better support for VPN DNS servers
Currently there is an issue with NetworkManager created WireGuard instances, that it will correctly add/update the VPN's DNS servers to /etc/valet-linux/dns-servers, but dnsmasq ignores the new DNS server. Adding this flag forces it to use the first entry of the dns-servers file which would be the wanted DNS server.